+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/handler/ExpirableLocalCacheHandler.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.subgrade.impl.handler;
 
+import com.dwarfeng.subgrade.sdk.exception.HandlerExceptionHelper;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import com.dwarfeng.subgrade.stack.handler.LocalCacheHandler;
 import org.slf4j.Logger;
@@ -98,7 +99,7 @@ public class ExpirableLocalCacheHandler<K, V> implements LocalCacheHandler<K, V>
                 lock.writeLock().unlock();
             }
         } catch (Exception e) {
-            throw new HandlerException(e);
+            throw HandlerExceptionHelper.parse(e);
         }
     }
 
@@ -149,7 +150,7 @@ public class ExpirableLocalCacheHandler<K, V> implements LocalCacheHandler<K, V>
                 lock.writeLock().unlock();
             }
         } catch (Exception e) {
-            throw new HandlerException(e);
+            throw HandlerExceptionHelper.parse(e);
         }
     }